+++ src/FuelClient.cc
@@ -370,6 +370,13 @@ Result FuelClient::DeleteModel(const ModelIdentifier &/*_id*/)
 //////////////////////////////////////////////////
 Result FuelClient::DownloadModel(const ModelIdentifier &_id)
 {
+  return this->DownloadModel(_id, {});
+}
+
+//////////////////////////////////////////////////
+Result FuelClient::DownloadModel(const ModelIdentifier &_id,
+    const std::vector<std::string> &_headers)
+{
   // Server config
   if (!_id.Server().Url().Valid() || _id.Server().Version().empty())
   {
@@ -387,7 +394,7 @@ Result FuelClient::DownloadModel(const ModelIdentifier &_id)
   ignition::fuel_tools::Rest rest;
   RestResponse resp;
   resp = rest.Request(HttpMethod::GET, _id.Server().Url().Str(),
-      _id.Server().Version(), route, {}, {}, "");
+      _id.Server().Version(), route, {}, _headers, "");
   if (resp.statusCode != 200)
   {
     ignerr << "Failed to download model." << std::endl