+++ src/ign.cc
@@ -438,7 +438,7 @@ extern "C" IGNITION_FUEL_TOOLS_VISIBLE int listWorlds(const char *_url,
 
 //////////////////////////////////////////////////
 extern "C" IGNITION_FUEL_TOOLS_VISIBLE int downloadUrl(const char *_url,
-    const char *_configFile)
+    const char *_configFile, const char *_header)
 {
   std::string urlStr{_url};
   ignition::common::URI url(urlStr);
@@ -479,7 +479,17 @@ extern "C" IGNITION_FUEL_TOOLS_VISIBLE int downloadUrl(const char *_url,
               << std::endl;
     }
 
-    auto result = client.DownloadModel(model);
+    int result = 0;
+    if (_header && strlen(_header) > 0)
+    {
+      std::vector<std::string> headers;
+      headers.push_back(_header);
+      result = client.DownloadModel(model, headers);
+    }
+    else
+    {
+      result = client.DownloadModel(model);
+    }
 
     if (!result)
     {